+++ drivers/pg/changestream/daemon.go
@@ -17,13 +17,10 @@ import (
 
 const (
 	// Limit batch sizes
-	BATCH_SIZE = 1_000
+	batchSize = 1_000
 )
 
 var (
-	// todo: why are the following properties safe to always ignore when computing hashes?
-	// todo: the common, ad, and azure packages where these string constants are defined are
-	// in ce repo. whats a good way to pull these constants in to dawgs?
 	ignoredPropertiesKeys = map[string]struct{}{
 		// common.ObjectID.String():      {},
 		// common.LastSeen.String():      {},
@@ -35,79 +32,256 @@ var (
 	}
 )
 
-type Log interface {
-	ResolveNodeChangeStatus(ctx context.Context, proposedChange *NodeChange) (ChangeStatus, error)
-	EvaluateNodeChange(proposedChange *NodeChange) (ChangeStatus, error)
-	LastEdgeChange(ctx context.Context, proposedChange *EdgeChange) (ChangeStatus, error)
-	CachedLastEdgeChange(proposedChange *EdgeChange) (ChangeStatus, error)
-	Submit(ctx context.Context, change Change) bool
+// todo: use golang-LRU cache
+// time multiple ingest runs with no cache, size 1_000, 100_000
+type ChangeCache struct {
+	data  map[string]ChangeStatus
+	mutex *sync.RWMutex
 }
 
-type Daemon struct {
-	// ingest writes to writerC.
-	writerC chan<- Change
-	// we continuously read changes from readerC to insert into pg
-	readerC <-chan Change
-
-	pgxPool          *pgxpool.Pool
-	kindMapper       pg.KindMapper
-	changeCacheLock  *sync.RWMutex
-	changeCache      map[string]ChangeStatus
-	State            *stateManager // todo: can we make this private?
-	nodeChangeBuffer []*NodeChange
-	edgeChangeBuffer []*EdgeChange
-}
-
-func NewDaemon(ctx context.Context, flags GetFlagByKeyer, pgxPool *pgxpool.Pool, kindMapper pg.KindMapper) *Daemon {
-	return &Daemon{
-		changeCache:     make(map[string]ChangeStatus),
-		changeCacheLock: &sync.RWMutex{},
-		State:           newStateManager(flags),
-		pgxPool:         pgxPool,
-		kindMapper:      kindMapper,
+func newChangeCache() ChangeCache {
+	return ChangeCache{
+		data:  make(map[string]ChangeStatus),
+		mutex: &sync.RWMutex{},
 	}
 }
 
-func (s *Daemon) PGX() *pgxpool.Pool {
-	return s.pgxPool
+func (s *ChangeCache) get(key string) (ChangeStatus, bool) {
+	s.mutex.RLock()
+	defer s.mutex.RUnlock()
+	val, ok := s.data[key]
+	return val, ok
 }
 
-func (s *Daemon) EvaluateNodeChange(proposedChange *NodeChange) (ChangeStatus, error) {
+func (s *ChangeCache) put(key string, value ChangeStatus) {
+	s.mutex.Lock()
+	defer s.mutex.Unlock()
+	s.data[key] = value
+}
+
+func (s *ChangeCache) evaluateNodeChange(proposedChange *NodeChange) (ChangeStatus, error) {
 	var (
-		lastChange  ChangeStatus
+		status      ChangeStatus
 		identityKey = proposedChange.IdentityKey()
 	)
 
 	if propertiesHash, err := proposedChange.Properties.Hash(ignoredPropertiesKeys); err != nil {
-		return lastChange, err
+		return status, err
+	} else if kindsHash, err := proposedChange.Kinds.Hash(); err != nil {
+		return status, err
 	} else {
 		// Track the properties hash and kind IDs
-		lastChange.PropertiesHash = propertiesHash
+		combinedHash := append(propertiesHash, kindsHash...)
+		status.PropertiesHash = combinedHash
 	}
 
-	if cachedChange, hasCachedChange := s.LastCachedChange(identityKey); hasCachedChange {
-		lastChange.Changed = !bytes.Equal(lastChange.PropertiesHash, cachedChange.PropertiesHash)
-		lastChange.Exists = true
+	if cachedChange, ok := s.get(identityKey); ok {
+		status.Changed = !bytes.Equal(status.PropertiesHash, cachedChange.PropertiesHash)
+		status.Exists = true
 	} else {
-		// If the change log is disabled then mark every non-cached lookup as changed
-		lastChange.Changed = !s.State.isEnabled()
+		// mark every non-cached lookup as changed
+		status.Changed = true
 	}
 
 	// Ensure this makes it into the cache before returning
-	s.PutCachedChange(identityKey, lastChange)
-	return lastChange, nil
+	s.put(identityKey, status)
+	return status, nil
 }
 
-func (s *Daemon) ResolveNodeChangeStatus(ctx context.Context, proposedChange *NodeChange) (ChangeStatus, error) {
-	lastChange, err := s.EvaluateNodeChange(proposedChange)
+type ChangeWriter struct {
+	PGX        *pgxpool.Pool
+	KindMapper pg.KindMapper
+}
+
+func newChangeWriter(pgxPool *pgxpool.Pool, kindMapper pg.KindMapper) ChangeWriter {
+	return ChangeWriter{
+		PGX:        pgxPool,
+		KindMapper: kindMapper,
+	}
+}
+
+func (s *ChangeWriter) flushNodeChanges(ctx context.Context, changes []*NodeChange) error {
+	// Early exit check for empty buffer flushes
+	if len(changes) == 0 {
+		return nil
+	}
+
+	var (
+		numChanges  = len(changes)
+		now         = time.Now()
+		copyColumns = []string{
+			"node_id",
+			"kind_ids",
+			"hash",
+			"change_type",
+			"created_at",
+		}
+	)
+
+	iterator := func(i int) ([]any, error) {
+		c := changes[i]
+
+		if mappedKindIDs, err := s.KindMapper.MapKinds(ctx, c.Kinds); err != nil {
+			return nil, fmt.Errorf("node kind ID mapping error: %w", err)
+		} else if hash, err := c.Hash(); err != nil {
+			return nil, err
+		} else {
+			rows := []any{
+				c.NodeID,
+				mappedKindIDs,
+				hash,
+				c.Type(),
+				now,
+			}
+
+			return rows, nil
+		}
+	}
+
+	if _, err := s.PGX.CopyFrom(ctx, pgx.Identifier{"node_change_stream"}, copyColumns, pgx.CopyFromSlice(numChanges, iterator)); err != nil {
+		return fmt.Errorf("change stream node change insert error: %v", err)
+	} else {
+		return nil
+	}
+}
+
+type loop struct {
+	State         *stateManager
+	ReaderC       <-chan Change
+	WriterC       chan<- Change
+	ChangeWriter  ChangeWriter
+	Cache         ChangeCache
+	FlushInterval time.Duration
+	NodeBuffer    []*NodeChange
+	BatchSize     int
+}
+
+func newLoop(ctx context.Context, state *stateManager, writer ChangeWriter, cache ChangeCache, batchSize int) loop {
+	writerC, readerC := channels.BufferedPipe[Change](ctx)
+
+	return loop{
+		State:         state,
+		ReaderC:       readerC,
+		WriterC:       writerC,
+		ChangeWriter:  writer,
+		Cache:         cache,
+		FlushInterval: 5 * time.Second,
+		NodeBuffer:    make([]*NodeChange, 0),
+		BatchSize:     batchSize,
+	}
+}
+
+func (s *loop) start(ctx context.Context) error {
+	ticker := time.NewTicker(s.FlushInterval)
+
+	defer func() {
+		close(s.WriterC)
+		ticker.Stop()
+		slog.InfoContext(ctx, "Shutting down change stream")
+	}()
+
+	// initialize the node_change_stream, edge_change_stream tables
+	if _, err := s.ChangeWriter.PGX.Exec(ctx, ASSERT_TABLE_SQL); err != nil {
+		return fmt.Errorf("failed asserting changelog tablespace: %w", err)
+	}
+
+	slog.InfoContext(ctx, "Starting change stream")
+
+	lastNodeWatermark := 0
+
+	for {
+		select {
+		case <-ctx.Done():
+			return nil
+
+			// todo: the following two cases are missing some logic around feature flag checking and watermark logic
+		case change := <-s.ReaderC:
+			if !s.State.isEnabled() {
+				continue
+			}
+
+			switch typed := change.(type) {
+			case *NodeChange:
+				s.NodeBuffer = append(s.NodeBuffer, typed)
+				if len(s.NodeBuffer) >= s.BatchSize {
+					// todo: error handling on flush?
+					if err := s.ChangeWriter.flushNodeChanges(ctx, s.NodeBuffer); err != nil {
+						slog.Warn(err.Error())
+					}
+					s.NodeBuffer = s.NodeBuffer[:0]
+				}
+			}
+
+		case <-ticker.C:
+			lastNodeWatermark = s.maybeFlush(ctx, lastNodeWatermark)
+		}
+	}
+}
+
+// maybeFlush checks if the node buffer has remained the same size over two ticks.
+// This implies inactivity—no new changes—and triggers a flush of any remaining changes.
+// The function returns the new watermark for tracking buffer growth across ticks.
+func (s *loop) maybeFlush(ctx context.Context, lastNodeWatermark int) int {
+	numNodeChanges := len(s.NodeBuffer)
+
+	hasNodeChangesToFlush := numNodeChanges > 0 && numNodeChanges == lastNodeWatermark
+
+	if !hasNodeChangesToFlush { // &&!hasEdgeChangesToFlush
+		// No eligible flush needed
+		lastNodeWatermark = numNodeChanges
+		return lastNodeWatermark
+	}
+
+	if hasNodeChangesToFlush {
+		if err := s.ChangeWriter.flushNodeChanges(ctx, s.NodeBuffer); err != nil {
+			slog.Warn(err.Error())
+		}
+		s.NodeBuffer = s.NodeBuffer[:0]
+	}
+
+	lastNodeWatermark = len(s.NodeBuffer)
+
+	return lastNodeWatermark
+}
+
+// todo: ChangeLog is the public export for this package to be consumed by bloodhound.
+type Changelog struct {
+	Cache  ChangeCache
+	Writer ChangeWriter
+	Loop   loop
+}
+
+func NewChangelogDaemon(ctx context.Context, flags GetFlagByKeyer, pgxPool *pgxpool.Pool, kindMapper pg.KindMapper, batchSize int) *Changelog {
+	cache := newChangeCache()
+	writer := newChangeWriter(pgxPool, kindMapper)
+	state := newStateManager(flags)
+	loop := newLoop(ctx, state, writer, cache, batchSize)
+
+	// prime the feature flag upon initalization
+	// because state is not a pointer value this doesn't actually update state
+	// so i changed it to a pointer
+	state.CheckFeatureFlag(ctx)
+
+	go loop.start(ctx)
+
+	return &Changelog{
+		Cache:  cache,
+		Writer: writer,
+		Loop:   loop,
+	}
+}
+
+func (s *Changelog) ResolveNodeChangeStatus(ctx context.Context, proposedChange *NodeChange) (ChangeStatus, error) {
+	lastChange, err := s.Cache.evaluateNodeChange(proposedChange)
 
 	if err != nil || lastChange.Exists {
 		return lastChange, err
 	}
 
-	if s.State.isEnabled() {
+	// todo: move state should be top level so no gross nesting
+	if s.Loop.State.isEnabled() {
 		var (
-			lastChangeRow = s.pgxPool.QueryRow(ctx, LAST_NODE_CHANGE_SQL, proposedChange.NodeID, lastChange.PropertiesHash)
+			lastChangeRow = s.Writer.PGX.QueryRow(ctx, LAST_NODE_CHANGE_SQL, proposedChange.NodeID, lastChange.PropertiesHash)
 			err           = lastChangeRow.Scan(&lastChange.Changed, &lastChange.Type)
 		)
 
@@ -125,191 +299,13 @@ func (s *Daemon) ResolveNodeChangeStatus(ctx context.Context, proposedChange *No
 		}
 
 		// Ensure this makes it into the cache before returning
-		s.PutCachedChange(proposedChange.IdentityKey(), lastChange)
+		s.Cache.put(proposedChange.IdentityKey(), lastChange)
 	}
 
 	return lastChange, nil
+
 }
 
-func (s *Daemon) Submit(ctx context.Context, change Change) bool {
-	return channels.Submit(ctx, s.writerC, change)
-}
-
-func (s *Daemon) LastCachedChange(cacheKey string) (ChangeStatus, bool) {
-	s.changeCacheLock.RLock()
-	defer s.changeCacheLock.RUnlock()
-
-	cachedChange, hasCachedChange := s.changeCache[cacheKey]
-	return cachedChange, hasCachedChange
-}
-
-func (s *Daemon) PutCachedChange(cacheKey string, cachedLookup ChangeStatus) {
-	s.changeCacheLock.Lock()
-	defer s.changeCacheLock.Unlock()
-
-	s.changeCache[cacheKey] = cachedLookup
-}
-
-func (s *Daemon) RunLoop(ctx context.Context) error {
-	// initialize the node_change_stream, edge_change_stream tables
-	if _, err := s.pgxPool.Exec(ctx, ASSERT_TABLE_SQL); err != nil {
-		return fmt.Errorf("failed asserting changelog tablespace: %w", err)
-	}
-
-	// todo: return to partitioning once initial implementation is working on one table
-	// if err := AssertChangelogPartition(ctx, s.pgxPool); err != nil {
-	// 	return fmt.Errorf("failed asserting changelog partition: %w", err)
-	// }
-
-	go s.runLoopBody(ctx, 5*time.Second)
-
-	return nil
-}
-
-func (s *Daemon) runLoopBody(ctx context.Context, flushInterval time.Duration) {
-	ticker := time.NewTicker(flushInterval)
-	defer func() {
-		s.pgxPool.Close()
-		close(s.writerC)
-		ticker.Stop()
-		slog.InfoContext(ctx, "Shutting down change stream")
-	}()
-
-	slog.InfoContext(ctx, "Starting change stream")
-
-	lastNodeWatermark := 0
-	// lastEdgeWatermark := 0
-
-	for {
-		now := time.Now()
-
-		select {
-		case <-ctx.Done():
-			return
-
-		case change := <-s.readerC:
-			s.handleIncomingChange(ctx, change, now)
-
-		case <-ticker.C:
-			lastNodeWatermark = s.maybeFlush(ctx, lastNodeWatermark)
-		}
-	}
-}
-
-func (s *Daemon) handleIncomingChange(ctx context.Context, change Change, now time.Time) {
-	if err := s.State.CheckFeatureFlag(ctx); err != nil {
-		slog.ErrorContext(ctx, fmt.Sprintf("feature flag check failed: %v", err))
-		return
-	}
-
-	if !s.State.isEnabled() {
-		return
-	}
-
-	// if err := s.State.checkChangelogPartitions(ctx, now, s.pgxPool); err != nil {
-	// 	slog.ErrorContext(ctx, fmt.Sprintf("partition check failed: %v", err))
-	// 	return
-	// }
-
-	s.QueueChange(ctx, now, change)
-}
-
-func (s *Daemon) QueueChange(ctx context.Context, now time.Time, nextChange Change) {
-	switch typedChange := nextChange.(type) {
-	case *NodeChange:
-		s.nodeChangeBuffer = append(s.nodeChangeBuffer, typedChange)
-
-		if len(s.nodeChangeBuffer) >= BATCH_SIZE {
-			if err := s.FlushNodeChanges(ctx); err != nil {
-				slog.ErrorContext(ctx, fmt.Sprintf("writing changes to change stream failed: %v", err))
-			}
-		}
-
-	case *EdgeChange:
-
-	default:
-		slog.ErrorContext(ctx, fmt.Sprintf("Unexpected change type: %T", nextChange))
-	}
-}
-
-func (s *Daemon) FlushNodeChanges(ctx context.Context) error {
-	// Early exit check for empty buffer flushes
-	if len(s.nodeChangeBuffer) == 0 {
-		return nil
-	}
-
-	var (
-		numChanges  = len(s.nodeChangeBuffer)
-		now         = time.Now()
-		copyColumns = []string{
-			"node_id",
-			"kind_ids",
-			"properties_hash",
-			"property_fields",
-			"change_type",
-			"created_at",
-		}
-
-		iteratorFunc = func(idx int) ([]any, error) {
-			nextNodeChange := s.nodeChangeBuffer[idx]
-
-			if mappedKindIDs, err := s.kindMapper.MapKinds(ctx, nextNodeChange.Kinds); err != nil {
-				return nil, fmt.Errorf("node kind ID mapping error: %w", err)
-			} else if propertiesHash, err := nextNodeChange.Properties.Hash(ignoredPropertiesKeys); err != nil {
-				return nil, fmt.Errorf("node properties hash error: %w", err)
-			} else {
-				rows := []any{
-					nextNodeChange.NodeID,
-					mappedKindIDs,
-					propertiesHash,
-					nextNodeChange.Properties.Keys(nil),
-					nextNodeChange.Type(),
-					now,
-				}
-
-				return rows, nil
-			}
-		}
-	)
-
-	slog.DebugContext(ctx, fmt.Sprintf("flushing %d node changes", numChanges))
-
-	// pgx.Identifier{partitionName}
-	if _, err := s.pgxPool.CopyFrom(ctx, pgx.Identifier{"node_change_stream"}, copyColumns, pgx.CopyFromSlice(numChanges, iteratorFunc)); err != nil {
-		slog.Info(fmt.Sprintf("change stream node change insert error: %v", err))
-	}
-
-	// TODO: Should the buffer clear be dependent on no errors?
-	s.nodeChangeBuffer = s.nodeChangeBuffer[:0]
-	return nil
-}
-
-// maybeFlush will only flush changes when the buffer has not increased in size in two ticks.
-// if the buffer is the same size that implies no new changes have arrived
-func (s *Daemon) maybeFlush(ctx context.Context, lastNodeWatermark int) int {
-	numNodeChanges := len(s.nodeChangeBuffer)
-
-	hasNodeChangesToFlush := numNodeChanges > 0 && numNodeChanges == lastNodeWatermark
-
-	if !hasNodeChangesToFlush { // &&!hasEdgeChangesToFlush
-		// No eligible flush needed
-		lastNodeWatermark = numNodeChanges
-		return lastNodeWatermark
-	}
-
-	// todo: add partitioning l8r
-	// if err := s.State.checkChangelogPartitions(ctx, now, s.pgxPool); err != nil {
-	// 	slog.ErrorContext(ctx, fmt.Sprintf("change log state check error: %v", err))
-	// 	return
-	// }
-
-	if hasNodeChangesToFlush {
-		if err := s.FlushNodeChanges(ctx); err != nil {
-			slog.ErrorContext(ctx, fmt.Sprintf("writing node changes to change stream failed: %v", err))
-		}
-	}
-
-	lastNodeWatermark = len(s.nodeChangeBuffer)
-
-	return lastNodeWatermark
+func (s *Changelog) Submit(ctx context.Context, change Change) bool {
+	return channels.Submit(ctx, s.Loop.WriterC, change)
 }
